requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
Automat==22.10.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
boto3==1.34.90
botocore==1.34.90
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.39.1
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.3.0
confection==0.1.4
ConfigArgParse==1.7
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
dlib==19.24.2
docopt==0.6.2
docutils==0.21.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
executing==2.0.1
fakeredis==2.22.0
fastjsonschema==2.19.1
fbmessenger==4.3.1
filelock==3.13.1
Flask==3.0.2
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gast==0.5.4
gevent==24.2.1
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.62.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
imageio==2.34.0
imagesize==1.4.1
imutils==0.5.4
incremental==22.10.0
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpickle==3.0.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.3
libclang==18.1.1
Mako==1.3.5
marisa-trie==1.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mathparse==0.1.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mysql-connector==2.2.9
mysql-connector-python==8.3.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
networkx==3.2.1
nltk==3.8.1
num2words==0.5.13
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.2.1
pandoc==2.3
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pefile==2023.2.7
pickleshare==0.7.5
Pint==0.23
platformdirs==4.2.0
plumbum==1.8.2
ply==3.11
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.25.3
pure-eval==0.2.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
pykwalify==1.6.0
PyMySQL==1.1.0
pynput==1.7.6
pyparsing==3.1.2
pypiwin32==223
pyreadline3==3.4.1
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
pyttsx3==2.90
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==26.0.2
redis==5.0.4
referencing==0.35.0
regex==2024.4.16
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==69.1.1
simplejson==3.19.2
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
smart-open==6.4.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.29
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
thinc==8.2.3
threadpoolctl==3.3.0
tifffile==2024.2.12
tinycss2==1.3.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
Tubes==0.2.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typer==0.9.4
typing==3.7.4.3
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
virtualenv==20.25.1
virtualenvwrapper-win==1.2.7
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
WTForms==3.1.2
yarg==0.1.9
yarl==1.9.4
zope.event==5.0
zope.interface==6.3
